+++ setuptools_rust/__init__.py
@@ -2,7 +2,6 @@ from .build import build_rust
 from .check import check_rust
 from .clean import clean_rust
 from .extension import Binding, RustExtension, Strip
-from .test import test_rust
 from .tomlgen import tomlgen_rust, find_rust_extensions
 from .version import version as __version__
 
@@ -15,6 +14,5 @@ __all__ = (
     "check_rust",
     "clean_rust",
     "find_rust_extensions",
-    "test_rust",
     "tomlgen_rust",
 )