+++ setuptools_rust/__init__.py
@@ -1,16 +1,6 @@
 from .build import build_rust
 from .clean import clean_rust
 from .extension import Binding, RustExtension, Strip
-from .tomlgen import tomlgen_rust, find_rust_extensions
 from .version import version as __version__
 
-
-__all__ = (
-    "Binding",
-    "RustExtension",
-    "Strip",
-    "build_rust",
-    "clean_rust",
-    "find_rust_extensions",
-    "tomlgen_rust",
-)
+__all__ = ("Binding", "RustExtension", "Strip", "build_rust", "clean_rust")